+++ bambi/interpret/types.py
@@ -5,9 +5,9 @@ from typing import Callable, NamedTuple
 
 import numpy as np
 import pandas as pd
-from arviz import InferenceData
 from pandas import DataFrame, Series
 from pandas.api.types import is_float_dtype, is_integer_dtype, is_numeric_dtype
+from xarray import DataTree
 
 from bambi.interpret.validate import (
     Values,
@@ -366,11 +366,11 @@ class SlopeVariable:
 
 
 class Result(NamedTuple):
-    """The result of an predictions, comparisons, or slopes computation.
+    """The result of a prediction, comparison, or slope computation.
 
     A `Result` contains a high-level summary dataframe and the original draws (samples)
     used to compute that summary.
     """
 
     summary: DataFrame
-    draws: InferenceData
+    draws: DataTree